fix: correct ar_params max_tokens and codec dtype for MOSS-TTS-Local inference
- end2end.py: ar_params max_tokens 18192→500 to prevent stage-0 generating
  128960 flat RVQ codes that exceed stage-1's max_model_len=18192.
  Each AR token produces 32 codes; floor(18192/32)=568 is the hard cap.

- moss_tts_decoder.py: add .float() after codec.to(device).eval() to ensure
  all codec parameters are float32, fixing 'Input type (float) and bias type
  (c10::BFloat16) should be the same' error during CAT codec decode.

+++ vllm_omni/model_executor/models/moss_tts/moss_tts_decoder.py
@@ -89,7 +89,7 @@ class CATCodecWorker:
             codec_path,
             trust_remote_code=True,
         )
-        self.codec = self.codec.to(self.device).eval()
+        self.codec = self.codec.to(self.device).eval().float()
 
         self.sample_rate: int = getattr(self.codec.config, "sampling_rate", 24_000)
         # Config exposes num_quantizers as a @property backed by quantizer_kwargs
